fix: Cast approved_by a integer en inventory-count
Corrige error "column approved_by is of type integer but expression is of type text"
al completar el conteo de inventario.

+++ src/app/api/market/pos/inventory-count/route.ts
@@ -504,7 +504,7 @@ export async function POST(request: NextRequest) {
           completed_at = NOW(),
           adjustment_operation_id = $2,
           auto_approved = $3,
-          approved_by = CASE WHEN $3 THEN $4 ELSE NULL END,
+          approved_by = CASE WHEN $3 THEN $4::integer ELSE NULL END,
           approved_at = CASE WHEN $3 THEN NOW() ELSE NULL END,
           updated_at = NOW()
         WHERE id = $5
